requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
environs==9.5.0
frozenlist==1.3.1
idna==3.3
magic-filter==1.0.8
marshmallow==3.17.0
multidict==6.0.2
packaging==21.3
pipedrive-python-lib==1.1.5
psycopg2-binary==2.9.3
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.11
yarl==1.8.1